requirements.txt
azure-storage-blob==12.19.0
azure-ai-documentintelligence==1.0.0
azure-core>=1.30.0
azure-search-documents==11.4.0
openai>=1.12.0
langchain>=0.1.0
python-dotenv==1.0.0
